Remove alpha and beta categories. Mark nodes in development by frames.
