+++ nodes/modifier_make/bmesh_analyzer_big.py
@@ -29,7 +29,7 @@ class SvBManalyzinNode(SverchCustomTreeNode, bpy.types.Node):
     bl_idname = 'SvBManalyzinNode'
     bl_label = 'BMesh Analyze In'
     bl_icon = 'OUTLINER_OB_EMPTY'
-    sv_icon = 'SV_ANALYZE_BMESH'
+    sv_icon = 'SV_ALPHA'  #'SV_ANALYZE_BMESH'
 
     mode1: BoolProperty(name='normal_update', default=True, update=updateNode)
