+++ nodes/modifier_make/bmesh_obj_in.py
@@ -28,7 +28,7 @@ class SvBMObjinputNode(SverchCustomTreeNode, bpy.types.Node):
     bl_idname = 'SvBMObjinputNode'
     bl_label = 'BMesh Obj In'
     bl_icon = 'OUTLINER_OB_EMPTY'
-    sv_icon = 'SV_OBJECT_IN_BMESH'
+    sv_icon = 'SV_ALPHA'  # 'SV_OBJECT_IN_BMESH'
 
     UseSKey: BoolProperty(name='with_shapekey', default=False, update=updateNode)
     keyIND: IntProperty(name='SHKey_ind', default=0, update=updateNode)
